# swagger/swagger.yaml
basePath: /v1
definitions:
  error_handler.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
  model.Permission:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  model.Role:
    properties:
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/model.Permission'
        type: array
    type: object
...
  model.User:
    properties:
      id:
        type: string
      role:
        $ref: '#/definitions/model.Role'
      role_id:
        type: string
      username:
        type: string
    type: object
  service.UpdateUserRequest:
    properties:
      password:
        type: string
      roleId:
        type: string
      username:
        type: string
    type: object
host: acc-api.jurmanovic.com
info:
  contact:
    name: ACC Server Manager Support
    url: https://github.com/FJurmanovic/acc-server-manager
  description: API for managing Assetto Corsa Competizione dedicated servers
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: ACC Server Manager API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and receive a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          properties:
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            properties:
              token:
                type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: User login
      tags:
      - Authentication
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get details of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Current user details
          schema:
            $ref: '#/definitions/model.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user details
      tags:
      - Authentication
  /auth/open-token:
    post:
      consumes:
      - application/json
      description: Generate an open token for a user
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            properties:
              token:
                type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      summary: Generate an open token
      tags:
      - Authentication
...
  /membership:
    get:
      consumes:
      - application/json
      description: Get a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - User Management
    post:
      consumes:
      - application/json
      description: Create a new user account with specified role
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          properties:
            password:
              type: string
            role:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Created user details
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - User Management
  /membership/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific user by ID
      parameters:
      - description: User ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: User successfully deleted
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - User Management
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific user
      parameters:
      - description: User ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - User Management
    put:
      consumes:
      - application/json
      description: Update user details by ID
      parameters:
      - description: User ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      - description: Updated user details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user details
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Invalid request body or ID format
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - User Management
  /membership/roles:
    get:
      consumes:
      - application/json
      description: Get a list of all available user roles
      produces:
      - application/json
      responses:
        "200":
          description: List of roles
          schema:
            items:
              $ref: '#/definitions/model.Role'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/error_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all roles
      tags:
      - User Management
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"